SlideController y UserController
