+++ src/main/java/com/alkemy/ong/controllers/UserController.java
@@ -1,7 +1,6 @@
 package com.alkemy.ong.controllers;
 
 import com.alkemy.ong.entities.User;
-import com.alkemy.ong.exceptions.UserServiceException;
 import com.alkemy.ong.pojos.input.UpdateUserDTO;
 import com.alkemy.ong.pojos.output.ListUserDTO;
 import com.alkemy.ong.services.RoleService;
@@ -12,12 +11,10 @@ import javax.validation.Valid;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.config.Configuration.AccessLevel;
 import org.modelmapper.convention.MatchingStrategies;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -62,22 +59,11 @@ public class UserController {
         modelMapper.map(userDto, user);
         userService.update(id, user);
     }
-    
+
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/{id}")
     public void delete(@PathVariable Long id){
         userService.delete(id);
     }
-    
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(UserServiceException.class)
-    public String UserServiceExceptionHandler(UserServiceException ex){
-        return ex.getMessage();
-    }
-    
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(NullPointerException.class)
-    public String NullPointerExceptionHandler(NullPointerException ex){
-        return ex.getMessage();
-    }
+
 }
